chp_matmul types fixed

+++ resilience-examples/checkpoint_matmul.c
@@ -41,7 +41,7 @@
 typedef struct resrv_carrier
 {
     int id;                     // ID of memory part that is being reserved
-    int *adr;                   // Address of memory part that is being reserved
+    uint64_t *adr;                   // Address of memory part that is being reserved
     int count;                  // number of data items that needs to be stored
     int pe_num;                 // the PE that asked for a reservation or checkpoint
     int is_symmetric;           // if this request is to checkpoint symmetric or private data
@@ -52,7 +52,7 @@ typedef struct check_carrier cpr_check_carrier;
 struct check_carrier
 {
     int id;                     // ID of memory part that is being checkpointed
-    int *adr;                   // Address of memory part that is being checkpointed
+    uint64_t *adr;                   // Address of memory part that is being checkpointed
     // TO DO: maybe if change from count to size=count*sizeof(data),
     // it can be generalized to all 
     int count;                  // number of data items that needs to be stored
@@ -381,7 +381,7 @@ void shmem_cpr_copy_carrier ( cpr_rsvr_carrier *frst, cpr_check_carrier *scnd )
     scnd -> is_symmetric = frst -> is_symmetric;
 }
 
-int shmem_cpr_is_reserved (int id, int *mem, int pe_num)
+int shmem_cpr_is_reserved (int id, uint64_t *mem, int pe_num)
 {
     switch(cpr_pe_type)
     {
